Segregate menus' logic
