+++ apps/meteor/client/components/message/toolbar/MessageToolbarActionMenu.tsx
@@ -1,0 +1,164 @@
+import { isE2EEMessage, type IMessage, type IRoom, type ISubscription } from '@rocket.chat/core-typings';
+import { useUniqueId } from '@rocket.chat/fuselage-hooks';
+import { GenericMenu, type GenericMenuItemProps } from '@rocket.chat/ui-client';
+import { useLayoutHiddenActions } from '@rocket.chat/ui-contexts';
+import { useQuery } from '@tanstack/react-query';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { useCopyAction } from './useCopyAction';
+import { useDeleteMessageAction } from './useDeleteMessageAction';
+import { useEditMessageAction } from './useEditMessageAction';
+import { useFollowMessageAction } from './useFollowMessageAction';
+import { useMarkAsUnreadMessageAction } from './useMarkAsUnreadMessageAction';
+import { useMessageActionAppsActionButtons } from './useMessageActionAppsActionButtons';
+import { useNewDiscussionMessageAction } from './useNewDiscussionMessageAction';
+import { usePermalinkAction } from './usePermalinkAction';
+import { usePinMessageAction } from './usePinMessageAction';
+import { useReadReceiptsDetailsAction } from './useReadReceiptsDetailsAction';
+import { useReplyInDMAction } from './useReplyInDMAction';
+import { useReportMessageAction } from './useReportMessageAction';
+import { useShowMessageReactionsAction } from './useShowMessageReactionsAction';
+import { useStarMessageAction } from './useStarMessageAction';
+import { useTranslateAction } from './useTranslateAction';
+import { useUnFollowMessageAction } from './useUnFollowMessageAction';
+import { useUnpinMessageAction } from './useUnpinMessageAction';
+import { useUnstarMessageAction } from './useUnstarMessageAction';
+import { useViewOriginalTranslationAction } from './useViewOriginalTranslationAction';
+import { useWebDAVMessageAction } from './useWebDAVMessageAction';
+import { MessageAction, type MessageActionContext } from '../../../../app/ui-utils/client/lib/MessageAction';
+import { roomsQueryKeys } from '../../../lib/queryKeys';
+
+type MessageActionSection = {
+	id: string;
+	title: string;
+	items: GenericMenuItemProps[];
+};
+
+type MessageToolbarActionMenuProps = {
+	message: IMessage;
+	context: MessageActionContext;
+	room: IRoom;
+	subscription: ISubscription | undefined;
+	onChangeMenuVisibility: (visible: boolean) => void;
+};
+
+const MessageToolbarActionMenu = ({ message, context, room, subscription, onChangeMenuVisibility }: MessageToolbarActionMenuProps) => {
+	// TODO: move this to another place
+	useWebDAVMessageAction(message, { subscription });
+	useNewDiscussionMessageAction(message, { room, subscription });
+	useUnpinMessageAction(message, { room, subscription });
+	usePinMessageAction(message, { room, subscription });
+	useStarMessageAction(message, { room });
+	useUnstarMessageAction(message, { room });
+	usePermalinkAction(message, { id: 'permalink-star', subscription, context: ['starred'], order: 10 });
+	usePermalinkAction(message, { id: 'permalink-pinned', subscription, context: ['pinned'], order: 5 });
+	usePermalinkAction(message, {
+		id: 'permalink',
+		subscription,
+		context: ['message', 'message-mobile', 'threads', 'federated', 'videoconf', 'videoconf-threads'],
+		type: 'duplication',
+		order: 5,
+	});
+	useFollowMessageAction(message, { room, context });
+	useUnFollowMessageAction(message, { room, context });
+	useMarkAsUnreadMessageAction(message, { room, subscription });
+	useTranslateAction(message, { room, subscription });
+	useViewOriginalTranslationAction(message, { room, subscription });
+	useReplyInDMAction(message, { room, subscription });
+	useCopyAction(message, { subscription });
+	useEditMessageAction(message, { room, subscription });
+	useDeleteMessageAction(message, { room, subscription });
+	useReportMessageAction(message, { room, subscription });
+	useShowMessageReactionsAction(message);
+	useReadReceiptsDetailsAction(message);
+
+	const hiddenActions = useLayoutHiddenActions().messageToolbox;
+	const { isSuccess, data } = useQuery({
+		queryKey: roomsQueryKeys.messageActionsWithParameters(room._id, message),
+		queryFn: async () => {
+			const menuItems = await MessageAction.getAll(context, 'menu');
+			return menuItems.filter((action) => !hiddenActions.includes(action.id));
+		},
+		keepPreviousData: true,
+	});
+
+	const actionButtonApps = useMessageActionAppsActionButtons(message, context);
+
+	const id = useUniqueId();
+	const { t } = useTranslation();
+
+	if (!isSuccess || data.length === 0) {
+		return null;
+	}
+
+	const options = [...data, ...(actionButtonApps.data ?? [])];
+
+	const isMessageEncrypted = isE2EEMessage(message);
+
+	const groupOptions = options
+		.map((option) => ({
+			variant: option.color === 'alert' ? 'danger' : '',
+			id: option.id,
+			icon: option.icon,
+			content: t(option.label),
+			onClick: option.action,
+			type: option.type,
+			...(typeof option.disabled === 'boolean' && { disabled: option.disabled }),
+			...(typeof option.disabled === 'boolean' &&
+				option.disabled && { tooltip: t('Action_not_available_encrypted_content', { action: t(option.label) }) }),
+		}))
+		.reduce(
+			(acc, option) => {
+				const group = option.type ? option.type : '';
+				const section = acc.find((section: { id: string }) => section.id === group);
+				if (section) {
+					section.items.push(option);
+					return acc;
+				}
+				const newSection = { id: group, title: group === 'apps' ? t('Apps') : '', items: [option] };
+				acc.push(newSection);
+
+				return acc;
+			},
+			[] as unknown as MessageActionSection[],
+		)
+		.map((section) => {
+			if (section.id !== 'apps') {
+				return section;
+			}
+
+			if (!isMessageEncrypted) {
+				return section;
+			}
+
+			return {
+				id: 'apps',
+				title: t('Apps'),
+				items: [
+					{
+						content: t('Unavailable'),
+						type: 'apps',
+						id,
+						disabled: true,
+						gap: false,
+						tooltip: t('Action_not_available_encrypted_content', { action: t('Apps') }),
+					},
+				],
+			};
+		});
+
+	return (
+		<GenericMenu
+			onOpenChange={onChangeMenuVisibility}
+			detached
+			title={t('More')}
+			data-qa-id='menu'
+			data-qa-type='message-action-menu-options'
+			sections={groupOptions}
+			placement='bottom-end'
+		/>
+	);
+};
+
+export default MessageToolbarActionMenu;
