+++ apps/meteor/client/components/message/toolbar/useMarkAsUnreadMessageAction.ts
@@ -1,6 +1,6 @@
 import { isOmnichannelRoom } from '@rocket.chat/core-typings';
-import type { ISubscription, IMessage, IRoom, IUser } from '@rocket.chat/core-typings';
-import { useRouter } from '@rocket.chat/ui-contexts';
+import type { ISubscription, IMessage, IRoom } from '@rocket.chat/core-typings';
+import { useRouter, useUser } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
 import { MessageAction } from '../../../../app/ui-utils/client';
@@ -8,8 +8,9 @@ import { useMarkAsUnreadMutation } from '../hooks/useMarkAsUnreadMutation';
 
 export const useMarkAsUnreadMessageAction = (
 	message: IMessage,
-	{ user, room, subscription }: { user: IUser | undefined; room: IRoom; subscription: ISubscription | undefined },
+	{ room, subscription }: { room: IRoom; subscription: ISubscription | undefined },
 ) => {
+	const user = useUser();
 	const { mutateAsync: markAsUnread } = useMarkAsUnreadMutation();
 
 	const router = useRouter();
