+++ apps/meteor/client/components/message/toolbar/useMessageActionAppsActionButtons.ts
@@ -5,12 +5,12 @@ import type { UseQueryResult } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { useAppActionButtons, getIdForActionButton } from './useAppActionButtons';
-import { useApplyButtonFilters } from './useApplyButtonFilters';
-import { UiKitTriggerTimeoutError } from '../../app/ui-message/client/UiKitTriggerTimeoutError';
-import type { MessageActionContext, MessageActionConfig } from '../../app/ui-utils/client/lib/MessageAction';
-import { Utilities } from '../../ee/lib/misc/Utilities';
-import { useUiKitActionManager } from '../uikit/hooks/useUiKitActionManager';
+import { UiKitTriggerTimeoutError } from '../../../../app/ui-message/client/UiKitTriggerTimeoutError';
+import type { MessageActionContext, MessageActionConfig } from '../../../../app/ui-utils/client/lib/MessageAction';
+import { Utilities } from '../../../../ee/lib/misc/Utilities';
+import { useAppActionButtons, getIdForActionButton } from '../../../hooks/useAppActionButtons';
+import { useApplyButtonFilters } from '../../../hooks/useApplyButtonFilters';
+import { useUiKitActionManager } from '../../../uikit/hooks/useUiKitActionManager';
 
 const filterActionsByContext = (context: string | undefined, action: IUIActionButton) => {
 	if (!context) {
